+++ sentry-rails/spec/sentry/rails/tracing/active_record_subscriber_spec.rb
@@ -28,7 +28,7 @@ RSpec.describe Sentry::Rails::Tracing::ActiveRecordSubscriber, :subscriber do
       expect(transaction[:spans].count).to eq(1)
 
       span = transaction[:spans][0]
-      expect(span[:op]).to eq("sql.active_record")
+      expect(span[:op]).to eq("db.sql.active_record")
       expect(span[:description]).to eq("SELECT \"posts\".* FROM \"posts\"")
       expect(span[:trace_id]).to eq(transaction.dig(:contexts, :trace, :trace_id))
     end