+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Bug Fixes
+
+- Allow overwriting of context values [#1724](https://github.com/getsentry/sentry-ruby/pull/1724)
+  - Fixes [#1722](https://github.com/getsentry/sentry-ruby/issues/1722)
+
 ## 5.1.0
 
 ### Features