+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 5.1.1
 
 ### Bug Fixes